+++ docs/examples/basic/usage.js
@@ -1,9 +1,0 @@
-const { FlowViewCanvas } = require('flow-view')
-
-const graph = require('./graph.json')
-
-const container = document.getElementById('drawing')
-
-const canvas = new FlowViewCanvas(container)
-
-canvas.loadGraph(graph)