+++ test/jsonSchema.js
@@ -1,10 +1,10 @@
 const assert = require('assert')
 const validator = require('is-my-json-valid')
 
-const schema = require('../docs/schema.json')
+const schema = require('../schema.json')
 
-const graph1 = require('../docs/examples/basic/graph.json')
-const graph2 = require('./graph.json')
+const graph1 = require('./graph1.json')
+const graph2 = require('./graph2.json')
 
 const validate = validator(schema)